eat(crypto): Add WPT tests and improve CryptoKey compatibility
