+++ modules/llrt_crypto/src/lib.rs
@@ -48,10 +48,11 @@ use rquickjs::{
     prelude::{Func, Rest},
     Class, Ctx, Error, Exception, Function, IntoJs, Null, Object, Result, Value,
 };
+pub use subtle::CryptoKey;
 use subtle::{
     subtle_decrypt, subtle_derive_bits, subtle_derive_key, subtle_digest, subtle_encrypt,
     subtle_export_key, subtle_generate_key, subtle_import_key, subtle_sign, subtle_unwrap_key,
-    subtle_verify, subtle_wrap_key, CryptoKey, SubtleCrypto,
+    subtle_verify, subtle_wrap_key, SubtleCrypto,
 };
 
 use self::{
