+++ tests/wpt/WebCryptoAPI.serialization.test.ts
@@ -1,0 +1,4 @@
+import { runSuite } from "./_harness-util.js";
+import { runTestDynamic } from "./WebCryptoAPI.harness.js";
+
+runSuite(import.meta.url, runTestDynamic);
